requirements.txt
discord.py==2.3.2
keyboard==0.13.5
psutil==5.9.8
pyautogui==0.9.54
python-dotenv==1.2.1
Requests==2.32.5
